+++ integration/drill/Dockerfile
@@ -1,12 +1,14 @@
 FROM java:openjdk-8-jdk
 # needs jdk to submit sql queries!?
 
+ENV DRILL_VERSION 1.15.0
+
 RUN	mkdir -p /opt/drill && \
-	wget -q -O - http://www-us.apache.org/dist/drill/drill-1.10.0/apache-drill-1.10.0.tar.gz | tar -zxvf  - -C /opt/drill
+	wget -q -O - http://www-us.apache.org/dist/drill/drill-${DRILL_VERSION}/apache-drill-${DRILL_VERSION}.tar.gz | tar -zxvf  - -C /opt/drill
 
 EXPOSE 8047
 
-ENV DRILL_HOME		/opt/drill/apache-drill-1.10.0
+ENV DRILL_HOME		/opt/drill/apache-drill-${DRILL_VERSION}
 ENV DRILL_LOG_DIR	${DRILL_HOME}/log/
 ENV DRILL_LOG_PREFIX	${DRILL_LOG_PATH}/drill
 
@@ -23,8 +25,8 @@ ENTRYPOINT ["java",\
 	"-Ddrill.exec.enable-epoll=false",\
 	"-XX:+CMSClassUnloadingEnabled",\
 	"-XX:+UseG1GC",\
-	"-Dlog.path=/opt/drill/apache-drill-1.10.0/log/drillbit.log",\
-	"-Dlog.query.path=/opt/drill/apache-drill-1.10.0/log/drillbit_queries.json",\
+	"-Dlog.path=/opt/drill/apache-drill-${DRILL_VERSION}/log/drillbit.log",\
+	"-Dlog.query.path=/opt/drill/apache-drill-${DRILL_VERSION}/log/drillbit_queries.json",\
 	"-cp",\
-	"/opt/drill/apache-drill-1.10.0/conf:/opt/drill/apache-drill-1.10.0/jars/*:/opt/drill/apache-drill-1.10.0/jars/ext/*:/opt/drill/apache-drill-1.10.0/jars/3rdparty/*:/opt/drill/apache-drill-1.10.0/jars/classb/*",\
+	"/opt/drill/apache-drill-${DRILL_VERSION}/conf:/opt/drill/apache-drill-${DRILL_VERSION}/jars/*:/opt/drill/apache-drill-${DRILL_VERSION}/jars/ext/*:/opt/drill/apache-drill-${DRILL_VERSION}/jars/3rdparty/*:/opt/drill/apache-drill-${DRILL_VERSION}/jars/classb/*",\
 	"org.apache.drill.exec.server.Drillbit"]
